Savio

+++ horario/views.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.models import Group
-from django.db.models import Q
+from django.contrib import messages
 from django.http import HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from ProjectWebII.utils import group_required
@@ -21,6 +21,7 @@ def add_disponibilidade(request):
             form = AddDisponibilidade(request.POST)
             if form.is_valid():
                 form.save()
+                messages.success(request, f'Disponibilidade cadastrado com sucesso!')
                 return redirect('add_disponibilidade')
         else:
             form = AddDisponibilidade()
@@ -29,7 +30,6 @@ def add_disponibilidade(request):
             'form': form,
             'profissionais': profissionais
         }
-
         return render(request, 'assets/static/crud_disponibilidade/add.html', context)
     else:
         return HttpResponse('erro!')
@@ -63,7 +63,7 @@ def remover_disponibilidade(request, dispo_id):
 
 def confirmar_remocao_disponibilidade(request, dispo_id):
     Disponibilidade.objects.get(pk=dispo_id).delete()
-
+    messages.error(request, f'Disponibilidade deletada com sucesso!')
     return redirect('add_disponibilidade')
 
 #CRUD de disponibilidade -- Fim
@@ -72,19 +72,26 @@ def confirmar_remocao_disponibilidade(request, dispo_id):
 @group_required(['Administrador', 'Profissional'], "/accounts/login/")
 def add_parametro(request):
     if request.user.is_authenticated:
-        parametro = Parametro.objects.all()
+        profissional_group = Group.objects.get(name='Profissional')
+        profissionais = profissional_group.user_set.all()
+        parametros = Parametro.objects.all()
+
         if request.method == 'POST':
+            print("teste")
             form = AddParametro(request.POST)
+            print("teste2")
             if form.is_valid():
+                print("test3")
                 form.save()
+                messages.success(request, f'Parâmetro adicionado com sucesso!')
                 return redirect('add_parametro')
         else:
             form = AddParametro()
         context = {
-            'parametros': parametro,
-            'form': form
+            'parametros': parametros,
+            'form': form,
+            'profissionais': profissionais
         }
-
         return render(request, 'assets/static/crud_parametro/add.html', context)
     else:
         return HttpResponse('erro!')
@@ -114,7 +121,7 @@ def remover_parametro(request, parametro_id):
 
 def confirmar_remocao_parametro(request, parametro_id):
     Parametro.objects.get(pk=parametro_id).delete()
-
+    messages.error(request, f'Parâmetro deletado com sucesso!')
     return redirect('add_parametro')
 
 #CRUD de parametros FIM
@@ -136,6 +143,8 @@ def gerar_horarios_disponiveis(request, profissional_id):
                     horario_disponivel=horario_disponivel
                 )
                 horarios.save()
+
+    messages.success(request, f'Horários de {disponibilidade.profissional} gerados com sucesso!')
     return redirect('listar_horarios')
 
 def calcular_horarios_disponiveis(disponibilidade):
@@ -187,8 +196,13 @@ def apagar_horarios_disponiveis(request, profissional_id):
     return render(request, 'assets/static/horarios/remove.html', context)
 
 def confirmar_remocao_horarios(request, profissional_id):
+    profissional = Usuario.objects.get(id=profissional_id)
     horarios = Horarios.objects.filter(disponibilidade__profissional=profissional_id)
-    horarios.delete()
+    if horarios.exists():
+        horarios.delete()
+        messages.error(request, f'Horários apagados com sucesso!')
+    else:
+        messages.warning(request, f'Horários de {profissional.username} já deletados!')
     return redirect('listar_horarios')
